llmAccountant/requirements.txt
PyPDF2 
streamlit 
google-generativeai
python-dotenv 
langchain 
chromadb
pdf2image